+++ modules/__tests__/proxy-test.js
@@ -1,6 +1,6 @@
 var expect = require('expect');
 var callApp = require('../utils/callApp');
-var forward = require('../forward');
+var proxy = require('../proxy');
 
 function ok() {
   return 'ok';
@@ -18,10 +18,10 @@ function returnFalse() {
   return false;
 }
 
-describe('mach.forward', function () {
+describe('mach.proxy', function () {
   describe('when no test function is given', function () {
     it('forwards the request', function () {
-      return callApp(forward(ok, target)).then(function (conn) {
+      return callApp(proxy(ok, target)).then(function (conn) {
         expect(conn.responseText).toEqual('target');
       });
     });
@@ -29,7 +29,7 @@ describe('mach.forward', function () {
 
   describe('a request that passes the test function', function () {
     it('is forwarded', function () {
-      return callApp(forward(ok, target, returnTrue)).then(function (conn) {
+      return callApp(proxy(ok, target, returnTrue)).then(function (conn) {
         expect(conn.responseText).toEqual('target');
       });
     });
@@ -37,11 +37,11 @@ describe('mach.forward', function () {
 
   describe('a request whose URL matches the test RegExp', function () {
     beforeEach(function () {
-      return callApp(forward(ok, target, /\/test-match$/), '/test-match');
+      return callApp(proxy(ok, target, /\/test-match$/), '/test-match');
     });
 
     it('is forwarded', function () {
-      return callApp(forward(ok, target, /\/test-match$/), '/test-match').then(function (conn) {
+      return callApp(proxy(ok, target, /\/test-match$/), '/test-match').then(function (conn) {
         expect(conn.responseText).toEqual('target');
       });
     });
@@ -49,7 +49,7 @@ describe('mach.forward', function () {
 
   describe('a request that does not pass the test function', function () {
     it('is not forwarded', function () {
-      return callApp(forward(ok, target, returnFalse), '/test-match').then(function (conn) {
+      return callApp(proxy(ok, target, returnFalse), '/test-match').then(function (conn) {
         expect(conn.responseText).toEqual('ok');
       });
     });