+++ modules/utils/createProxy.js
@@ -4,12 +4,12 @@ var proxyRequest = require('./proxyRequest');
 var MACH_VERSION = require('../version');
 
 /**
- * A mach.proxy is a function that is used to proxy a connection to
- * a different location.
+ * A proxy is a function that is used to forward a request to
+ * a different location and return the response.
  *
  * This function is part of the low-level API and can generally be
  * used more conveniently either through the client methods or the
- * mach.forward middleware.
+ * mach.proxy middleware.
  */
 function createProxy(location) {
   if (!(location instanceof Location))